+++ CHANGELOG.md
@@ -35,6 +35,7 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Update `FavoritesService::fetchAllFavorites` with peerId parameter.
 -   Add `FavoritesService::checkIsPeerSettedAsFavorite` API that can be used to check if a given peerId is in user's favorites.
 -   Add `FavoritesService::getFavoriteById` API that can be used to retrieve a specific user's favorite by Id. 
+-   Add `FavoritesService::getAllUserFavoriteList` API that can be used to retrieve the list of user's favorites. 
 
 ## [2.20.0] - 2023-02-24
 -   Add `PresenceService::setApplyMsTeamsPresenceSettings` This api allows to activate the exchange of presence of the connected user between rainbow and MS Teams on UI side.